requirement.txt
alabaster==0.7.12
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bokeh==2.3.2
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
conda==4.3.16
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
docutils==0.16
entrypoints==0.3
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
funcy==1.16
future==0.18.2
gensim==3.8.3
greenlet==1.1.0
idna==2.10
imagesize==1.2.0
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.2
notebook==6.4.0
numexpr==2.7.3
numpy==1.21.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
pkg==0.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.9.0
pyLDAvis==3.3.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-magic==0.4.24
pytz==2021.1
pyvis==0.1.9
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
smart-open==5.1.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.20
statsmodels==0.12.2
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tomotopy==0.12.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whiteboard==0.1.1
widgetsnbextension==3.5.1
